+++ bases/BaseController.d.ts
@@ -4,5 +4,4 @@ export default class BaseController {
     parse(request: Bun.BunRequest): Promise<Record<string, any>>;
     get response(): typeof Response;
     validate(validator: VineValidator<any, Record<string, any> | undefined>, body: Record<string, any>): Promise<any>;
-    private parseForm;
 }