+++ bases/BaseController.js
@@ -1,4 +1,5 @@
-import { defineValue, isEmpty, isNotEmpty } from "@bejibun/utils";
+import { defineValue, isNotEmpty } from "@bejibun/utils";
+import { default as Bobject } from "@bejibun/utils/facades/Object";
 import ValidatorException from "../exceptions/ValidatorException";
 import Response from "../facades/Response";
 export default class BaseController {
@@ -7,7 +8,7 @@ export default class BaseController {
         const formData = new FormData();
         try {
             if (contentType.includes("application/json"))
-                return await request.json();
+                return Bobject.serialize(await request.json());
             for (const [key, value] of Object.entries(request.params)) {
                 formData.append(key, value);
             }
@@ -29,7 +30,7 @@ export default class BaseController {
         catch {
             // do nothing
         }
-        return this.parseForm(formData);
+        return Bobject.parseFormData(formData);
     }
     get response() {
         return Response;
@@ -47,52 +48,5 @@ export default class BaseController {
                 message = defineValue(error?.message, defaultMessage);
             throw new ValidatorException(message);
         }
-    }
-    parseForm(formData) {
-        const result = {};
-        for (const [key, value] of formData.entries()) {
-            const keys = key.replace(/]/g, "").split("[");
-            let current = result;
-            for (let i = 0; i < keys.length; i++) {
-                const part = keys[i];
-                const nextPart = keys[i + 1];
-                if (i === keys.length - 1) {
-                    let convertedValue;
-                    if (value instanceof File) {
-                        convertedValue = value;
-                    }
-                    else if (value.trim() === "" || value === "null" || value === "undefined") {
-                        convertedValue = null;
-                    }
-                    else if (Number.isNaN(value)) {
-                        convertedValue = Number(value);
-                    }
-                    else if (value === "true" || value === "false") {
-                        convertedValue = value === "true";
-                    }
-                    else {
-                        try {
-                            convertedValue = JSON.parse(value);
-                        }
-                        catch {
-                            convertedValue = value;
-                        }
-                    }
-                    if (current[part] === undefined)
-                        current[part] = convertedValue;
-                    else if (Array.isArray(current[part]))
-                        current[part].push(convertedValue);
-                    else
-                        continue;
-                }
-                else {
-                    const isArrayIndex = /^\d+$/.test(nextPart);
-                    if (isEmpty(current[part]))
-                        current[part] = isArrayIndex ? [] : {};
-                    current = current[part];
-                }
-            }
-        }
-        return result;
     }
 }